Split landing into smaller components
