+++ web/common/routes.js
@@ -1,15 +1,15 @@
 import React from "react";
-import { Landing } from "../landing/landing";
-import Login from "../landing/login";
-import Signup from "../landing/signup/root";
+import { Landing } from "../landing/root";
+import Login from "../login/login";
+import Signup from "../signup/root";
 import Stats from "../landing/stats";
 import Home from "../home/AccountInfo";
-import { Invite } from "../landing/invite";
+import { Invite } from "../signup/invite";
 import { RedeemInvite } from "../home/RedeemInvite";
-import { FranceConnectCallback } from "../landing/signup/FranceConnectCallback";
-import { Logout } from "../landing/logout";
-import { ActivateEmail } from "../landing/signup/ActivateEmail";
-import { RequestResetPassword, ResetPassword } from "../landing/ResetPassword";
+import { FranceConnectCallback } from "../signup/FranceConnectCallback";
+import { Logout } from "../login/logout";
+import { ActivateEmail } from "../signup/ActivateEmail";
+import { RequestResetPassword, ResetPassword } from "../login/ResetPassword";
 import { CGU } from "../landing/cgu";
 import { UserRead } from "../control/UserRead";
 import { XlsxVerifier } from "../control/VerifyXlsxSignature";
@@ -129,7 +129,7 @@ export const ROUTES = [
     path: "/oauth/authorize",
     label: "OAuth",
     accessible: () => true,
-    component: React.lazy(() => import("../landing/oauth/root")),
+    component: React.lazy(() => import("../oauth/root")),
     menuItemFilter: () => false
   },
   {
