+++ web/oauth/root.js
@@ -10,7 +10,7 @@ import {
   useLocation,
   useHistory
 } from "react-router-dom";
-import { Header } from "../../common/Header";
+import { Header } from "../common/Header";
 import { API_HOST, useApi } from "common/utils/api";
 import { ConfirmUser } from "./ConfirmUser";
 import { useLoadingScreen } from "common/utils/loading";
